GetEnv and FileUpload

+++ GZ/HandleCGI/UploadFile.py
@@ -1,24 +1,9 @@
-import sys
 import streamlit as st
-import os
-
-"""
-
-env_dict = dict(os.environ)
-
-    Body> Stdinput
-
-
-
-"""
-
 
 st.write("File Uploader")
 
 content = st.file_uploader("Upload file")
 
-
-
-if content is not None:
+if content is not None and not os.path.isfile("userinput.txt"):
     with open("userinput.txt", "w+") as file:
         file.write(content)
